+++ packages/css-in-js/package.json
@@ -64,6 +64,7 @@
     "@emotion/hash": "^0.8.0",
     "@rocket.chat/css-supports": "workspace:packages/css-supports",
     "@rocket.chat/memo": "^0.28.0",
+    "@rocket.chat/stylis-logical-props-middleware": "workspace:packages/stylis-logical-props-middleware",
     "stylis": "^4.0.10"
   }
 }
