+++ packages/stylis-logical-props-middleware/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: '@rocket.chat/eslint-config-alt/typescript',
+  env: {
+    jest: true,
+  },
+};
