+++ packages/stylis-logical-props-middleware/.gitignore
@@ -1,0 +1,1 @@
+/dist
