+++ packages/stylis-logical-props-middleware/.prettierignore
@@ -1,0 +1,2 @@
+/node_modules
+/dist
