+++ packages/stylis-logical-props-middleware/.prettierrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@rocket.chat/prettier-config/fuselage');
