+++ packages/stylis-logical-props-middleware/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  preset: 'ts-jest',
+  errorOnDeprecated: true,
+  testMatch: ['<rootDir>/src/**/*.spec.ts'],
+  globals: {
+    'ts-jest': {
+      tsconfig: {
+        noUnusedLocals: false,
+        noUnusedParameters: false,
+      },
+    },
+  },
+};
