+++ packages/stylis-logical-props-middleware/src/index.ts
@@ -1,0 +1,4 @@
+import { createLogicalPropertiesMiddleware } from './middleware';
+
+export { createLogicalPropertiesMiddleware } from './middleware';
+export default createLogicalPropertiesMiddleware();
