+++ packages/stylis-logical-props-middleware/src/operations.ts
@@ -1,57 +1,13 @@
-import { cssSupports } from '@rocket.chat/css-supports';
-import {
-  DECLARATION,
-  Element,
-  Middleware,
-  node,
-  RULESET,
-  serialize,
-} from 'stylis';
+import { attachDeclaration, RuleSet, Declaration } from './elements';
 
-type RuleSet = Element & {
-  type: typeof RULESET;
-  children: Element[];
-  props: string[];
-};
-
-type Declaration = Element & {
-  type: typeof DECLARATION;
-  props: string;
-  children: string;
-};
-
-type Operation = (
+export type Operation = (
   value: Declaration['children'],
   ruleSet: Readonly<RuleSet>,
   ltrRuleSet: Readonly<RuleSet>,
   rtlRuleSet: Readonly<RuleSet>
 ) => void;
 
-const isRuleSet = (element: Element): element is RuleSet =>
-  element.type === RULESET;
-
-const isDeclaration = (element: Element): element is Declaration =>
-  element.type === DECLARATION;
-
-const attachDeclaration = (
-  property: string,
-  value: string,
-  ruleSet: RuleSet
-): void => {
-  const declaration = node(
-    `${property}:${value};`,
-    ruleSet,
-    ruleSet,
-    DECLARATION,
-    property as unknown as string[],
-    value as unknown as Element[],
-    property.length
-  );
-
-  ruleSet.children.push(declaration);
-};
-
-const compileOperations = ({
+export const compileOperations = ({
   isPropertySupported,
   isPropertyValueSupported,
 }: {
@@ -278,70 +234,3 @@ const compileOperations = ({
 
   return ops;
 };
-
-export const createLogicalPropertiesMiddleware = ({
-  isPropertySupported = (property: string): boolean =>
-    cssSupports(`${property}:inherit`),
-  isPropertyValueSupported = (property: string, value: string): boolean =>
-    cssSupports(`${property}:${value}`),
-}: {
-  isPropertySupported?: (property: string) => boolean;
-  isPropertyValueSupported?: (property: string, value: string) => boolean;
-} = {}): Middleware => {
-  const ops = compileOperations({
-    isPropertySupported,
-    isPropertyValueSupported,
-  });
-
-  return (ruleSet: RuleSet, _, __, callback): undefined | string => {
-    if (!isRuleSet(ruleSet) || ruleSet.root !== null) {
-      return undefined;
-    }
-
-    const ltrRuleSet = node(
-      ruleSet.props
-        .map((selector) => `html:not([dir=rtl]) ${selector}`)
-        .join(','),
-      undefined as unknown as Element,
-      undefined as unknown as Element,
-      RULESET,
-      ruleSet.props.map((selector) => `html:not([dir=rtl]) ${selector}`),
-      [],
-      0
-    ) as RuleSet;
-
-    const rtlRuleSet = node(
-      ruleSet.props.map((selector) => `[dir=rtl] ${selector}`).join(','),
-      undefined as unknown as Element,
-      undefined as unknown as Element,
-      RULESET,
-      ruleSet.props.map((selector) => `[dir=rtl] ${selector}`),
-      [],
-      0
-    ) as RuleSet;
-
-    const rules = ruleSet.children;
-    ruleSet.children = [];
-    ruleSet.return = '';
-
-    for (const rule of rules as Element[]) {
-      if (!isDeclaration(rule)) {
-        ruleSet.children.push(rule);
-        continue;
-      }
-
-      const op = ops.get(rule.props);
-
-      if (op) {
-        op(rule.children, ruleSet, ltrRuleSet, rtlRuleSet);
-        continue;
-      }
-
-      attachDeclaration(rule.props, rule.children, ruleSet);
-    }
-
-    return serialize([ltrRuleSet, rtlRuleSet], callback);
-  };
-};
-
-export const logicalPropertiesMiddleware = createLogicalPropertiesMiddleware();
