+++ packages/stylis-logical-props-middleware/tsconfig-esm.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["**/*.spec.ts"]
+}
