+++ packages/stylis-logical-props-middleware/typedoc.json
@@ -1,0 +1,6 @@
+{
+  "entryPoints": [
+    "src/index.ts"
+  ],
+  "out": "../../static/stylis-logical-props-middleware"
+}
